==> notebook/0_autism_terms.ipynb <==
import pandas as pd
import numpy as np

# load DisGenet curated raw data
df = pd.read_csv("../raw_data/curated_gene_disease_associations.tsv", sep='\t')

# Extract columns for disorders
diseases = df[['diseaseId', 'diseaseName']]

diseases = diseases.drop_duplicates()

# Find disorders for autism

## 1. Polygenetic autism spetcrum

poly_autism_terms = {}

### Autism spectrum

autism_spectrum = diseases[diseases['diseaseName'].str.lower().str.contains('autis')]
autism_spectrum.values

Delete syndromatic autism or ambiguous disorders
- C1854416: MACROCEPHALY/AUTISM SYNDROME
- C4013764: MENTAL RETARDATION WITH LANGUAGE IMPAIRMENT AND WITH OR WITHOUT AUTISTIC FEATURES

autism_spectrum = autism_spectrum[autism_spectrum['diseaseId'] != "C1854416"]
autism_spectrum = autism_spectrum[autism_spectrum['diseaseId'] != "C4013764"]

poly_autism_terms['Autism Spectrum Disorder'] = autism_spectrum

### Asperger's syndrome

asperger = diseases[diseases['diseaseName'].str.lower().str.contains('asper')]
asperger.values

# delete irrelevant diagnosis
asperger = asperger[asperger['diseaseName'] != 'Aspergillosis']
asperger.values

poly_autism_terms['Asperger Syndrome'] = asperger

### Find pervasive developmental disorder

pervasive_evelopmental_disorders = diseases[diseases['diseaseName'].str.lower().str.contains('development') & diseases['diseaseName'].str.lower().str.contains('pervasive')]
pervasive_evelopmental_disorders.values

poly_autism_terms['Pervasive Development Disorder'] = pervasive_evelopmental_disorders

### Combine tables

for key, df in poly_autism_terms.items():
    df['autism_subtype'] = key

autism_terms = pd.concat(poly_autism_terms.values(), 
                         ignore_index=True)
autism_terms.values

autism_terms['autism_subtype'] = 'Autism Spectrum Disorder'
autism_terms

autism_terms['autism_subtype_broad'] = 'Non-Syndromic Autism'

## 2. Syndromic ASD

autism_syndromes = {}

### Childhood disintegrative disorder, aka, Heller syndrome

hellers_syndrome = diseases[
    (diseases['diseaseName'].str.lower().str.contains('disintegrat') & diseases['diseaseName'].str.lower().str.contains('disorder')) |
    (diseases['diseaseName'].str.lower().str.contains('heller')) |
    (diseases['diseaseName'].str.lower().str.contains('cdd'))
]

hellers_syndrome.values

autism_syndromes['Heller Syndrome'] = hellers_syndrome

### Rett Syndrome - finally decided not to add

rett_syndrome = diseases[
    (diseases['diseaseName'].str.lower().str.contains('rett') & diseases['diseaseName'].str.lower().str.contains('syndrome')) &
    ~diseases['diseaseName'].isin(['Gilles de la Tourette syndrome', 'Baraitser Brett Piesowicz syndrome'])
]

rett_syndrome.values

autism_syndromes["Rett Syndrome"] = rett_syndrome

### Fragile X syndrome

fragile_x = diseases[
    (diseases['diseaseName'].str.lower().str.contains('fragile') & diseases['diseaseName'].str.lower().str.contains('x'))
]

fragile_x.values

autism_syndromes['Fragile X Syndrome'] = fragile_x

### MECP2 duplication syndrome

mecp2 = diseases[
    diseases['diseaseName'].str.lower().str.contains('mecp2')
]

mecp2.values

autism_syndromes['MECP2 duplication Syndrome'] = mecp2

### Tuberous sclerosis complex

ts = diseases[
    (diseases['diseaseName'].str.lower().str.contains('tuberous')) & (diseases['diseaseName'].str.lower().str.contains('sclerosis'))
]

ts.values

autism_syndromes['Tuberous Sclerosis Complex'] = ts

### Angelman's syndrome

angelman = diseases[
    diseases['diseaseName'].str.lower().str.contains('angelman')
]

angelman.values

autism_syndromes['Angelman Syndrome']=angelman

### phelan-mcdermid syndrome: not found

pm = diseases[
    (diseases['diseaseName'].str.lower().str.contains('phelan')) & (diseases['diseaseName'].str.lower().str.contains('pms'))
]

pm.values

autism_syndromes['Phelan-Mcdermid Syndrome']=pm

### Timothy syndrome

timothy = diseases[
    diseases['diseaseName'].str.lower().str.contains('timothy')
]

timothy.values

autism_syndromes['Timothy Syndrome']=timothy

### Smith-lemli-opitz syndrome

slo = diseases[
    (diseases['diseaseName'].str.lower().str.contains('smith')) & (diseases['diseaseName'].str.lower().str.contains('lemli'))
]

slo.values

autism_syndromes['Smith-Lemli-Opitz Syndrome']=slo

### neurofibromatosis 

neurofibromatosis = diseases[
    diseases['diseaseName'].str.lower().str.contains('neurofibromatosis')
]

neurofibromatosis.values

Delete ambiguous disorder, as it should be labeled as Nooan syndrome in this dataset
- 'C2931482': 'Neurofibromatosis-Noonan syndrome'

neurofibromatosis = neurofibromatosis[neurofibromatosis['diseaseId'] != 'C2931482']
neurofibromatosis.values

autism_syndromes['Neurofibromatosis']=neurofibromatosis

### PTEN hamartoma tumor syndrome

hamartoma = diseases[
    (diseases['diseaseName'].str.lower().str.contains('hamartoma')) & diseases['diseaseName'].str.lower().str.contains('syndrome')
]

hamartoma.values

autism_syndromes['Hamartoma tumor Syndrome']=hamartoma

### Down syndrome

down = diseases[
    (diseases['diseaseName'].str.lower().str.contains('down')) & diseases['diseaseName'].str.lower().str.contains('syndrome')
]

down.values

autism_syndromes['Down Syndrome']=down

### Cohen's syndrome

cohen = diseases[
    (
        (diseases['diseaseName'].str.lower().str.contains('cohen')) | 
        (diseases['diseaseName'].str.lower().str.contains('pepper')) |
        (diseases['diseaseName'].str.lower().str.contains('cervenka'))
    ) & 
    (
        diseases['diseaseName'].str.lower().str.contains('syndrome')
    )

]

cohen.values

autism_syndromes['Cohen Syndrome']=cohen

### Cornelia de lange syndrome

cornelia = diseases[
    (
        (diseases['diseaseName'].str.lower().str.contains('cornelia')) | 
        (diseases['diseaseName'].str.lower().str.contains('brachmann')) |
        (diseases['diseaseName'].str.lower().str.contains('bushy'))
    ) & 
    (
        (diseases['diseaseName'].str.lower().str.contains('syndrome')) |
        (diseases['diseaseName'].str.lower().str.contains('dwarf'))
    )
]


cornelia.values

autism_syndromes['Cornelia de Lange Syndrome']=cornelia

### Charge syndrome

charge = diseases[
    (diseases['diseaseName'].str.lower().str.contains('charge')) & 
    ~(diseases['diseaseName'].str.lower().str.contains('discharge')) 
]


charge.values

autism_syndromes['Charge Syndrome']=charge

### Noonan syndrome

noonan = diseases[
    (diseases['diseaseName'].str.lower().str.contains('noonan')) & 
    (diseases['diseaseName'].str.lower().str.contains('syndrome')) 
]

noonan.values

autism_syndromes['Noonan Syndrome']=noonan

Find Williams syndrome

william = diseases[
    (
        (diseases['diseaseName'].str.lower().str.contains('william')) & 
        (diseases['diseaseName'].str.lower().str.contains('syndrome'))
    ) | (
        (diseases['diseaseName'].str.lower().str.contains('7q11')) & 
        (diseases['diseaseName'].str.lower().str.contains('deletion'))
    )
]

william.values

autism_syndromes['William Syndrome']=william

### DiGeorge or 22q11.2 deletion syndrome

digeorge = diseases[
    (
        (diseases['diseaseName'].str.lower().str.contains('digeorge')) & 
        (diseases['diseaseName'].str.lower().str.contains('syndrome'))
    ) | (
        (diseases['diseaseName'].str.lower().str.contains('22q11')) & 
        (diseases['diseaseName'].str.lower().str.contains('deletion'))
    )
]

digeorge.values

autism_syndromes['Digeorge Syndrome']=digeorge

### Macrocephaly/autism syndrome 

macrocephaly = diseases[diseases['diseaseName'].str.upper().str.contains('MACROCEPHALY/AUTISM SYNDROME')]

macrocephaly.values

autism_syndromes['Macrocephaly/Autism Syndrome ']=macrocephaly

#### Combine syndromic autism

for key, df in autism_syndromes.items():
    df['autism_subtype'] = key

autism_syndromes['Fragile X Syndrome']

autism_syndromes_df = pd.concat(autism_syndromes.values(), axis=0, ignore_index=True)

autism_syndromes_df['autism_subtype_broad'] = 'Syndromic Autism'
autism_syndromes_df

# Combine & Export to csv

df = pd.concat([autism_terms, autism_syndromes_df], axis=0)

df.head()

df.tail()

df.to_csv('processed_data/autism_terms.csv', sep='\t')
