==> notebook/1_group_autism_subtypes.ipynb <==
import pandas as pd
import numpy as np

## Load data

# load DisGenet curated raw data
df = pd.read_csv("../raw_data/curated_gene_disease_associations.tsv", sep='\t')
df.head(1)

# Create diseases 
diseases = df[['diseaseId', 'diseaseName', 'diseaseType', 'diseaseClass', 'diseaseSemanticType']].drop_duplicates()
diseases.head()

# load autism terms
autisms = pd.read_csv('processed_data/autism_terms.csv', sep='\t').iloc[:,1:]
autisms.head()

autisms['diseaseId'] = autisms['diseaseId'].astype(str)
diseases['diseaseId'] = diseases['diseaseId'].astype(str)

autisms_df = autisms.merge(
    diseases[['diseaseId', 'diseaseType', 'diseaseClass', 'diseaseSemanticType']],
    on="diseaseId",
    how='left')

autisms_df.tail()

## Create col: disease ids, name, autism_subtype

autism_types_names = pd.Series(autisms_df['autism_subtype'].unique())

autism_types_new_id = "A" + pd.Series([str(i).zfill(2) for i in range(0, len(autism_types_names))])
autism_new_ids = pd.concat([autism_types_names, autism_types_new_id], axis=1)
autism_new_ids = autism_new_ids.rename(columns={0: 'diseaseName', 1:"diseaseId"})
autism_new_ids

# Create new columns
autism_new_ids.loc[0, 'autism_subtype_broad'] = 'Non-Syndromic Autism'
autism_new_ids.loc[1:, 'autism_subtype_broad'] = 'Syndromic Autism'
autism_new_ids.head(2)

## Create col: disease type


autisms_df[['autism_subtype', 'diseaseType']].drop_duplicates()

autism_new_ids['diseaseType'] = 'disease'
autism_new_ids.head(2)

## Create col: diseaseClass

autism_classes = autisms_df[['autism_subtype', 'diseaseClass']].drop_duplicates()
autism_classes.head()

for i, row in autism_new_ids.iterrows(): 
    df_temp = autism_classes[autism_classes['autism_subtype']==row['diseaseName']]
    list_temp = []
    for _, row_2 in df_temp.iterrows():
        if row_2['diseaseClass'] is not np.nan:
            classes = row_2['diseaseClass']
            if ";" in classes:
                list_temp += (classes.split(';'))
            else:
                list_temp.append(classes)
    list_temp = sorted(list(set(list_temp)))
    print(list_temp)
    autism_new_ids.loc[i, 'diseaseClass'] = ";".join(list_temp)

### Create col: 'diseaseSemanticType'

autism_semantictypes = autisms_df[['autism_subtype', 'diseaseSemanticType']].drop_duplicates()
autism_semantictypes = autism_semantictypes.reset_index().iloc[:, 1:]
autism_semantictypes

def relabel(autism_type, semantic_type):
    autism_semantictypes.loc[autism_semantictypes['autism_subtype']==autism_type, 'diseaseSemanticType'] = semantic_type

relabel('Autism Spectrum Disorder', 'Mental or Behavioral Dysfunction')
relabel('Tuberous Sclerosis Complex', 'Disease or Syndrome')
relabel('Neurofibromatosis', 'Disease or Syndrome')
relabel('Down Syndrome', 'Disease or Syndrome')
relabel('Cohen Syndrome', 'Congenital Abnormality')

autism_semantictypes = autism_semantictypes.drop_duplicates().reset_index(drop=True)
autism_semantictypes

autism_new_ids = autism_new_ids.merge(autism_semantictypes, left_on='diseaseName', right_on='autism_subtype', how='inner')
autism_new_ids = autism_new_ids.drop('autism_subtype', axis=1)

autism_new_ids

## Group autisms in raw data 

autism_new_ids.head(1)

old_autism_keys = autisms[['diseaseId', 'autism_subtype']]
old_autism_keys = old_autism_keys.rename(columns={'diseaseId': 'diseaseId_old', 'autism_subtype': 'diseaseName'})

autism_new_ids = autism_new_ids.merge(old_autism_keys, on='diseaseName', how='left')

autism_new_ids.head(1)

autism_new_ids = autism_new_ids.rename(columns={
    'diseaseName': 'new_diseaseName',
    'diseaseId': 'new_diseaseId',
    'diseaseType': 'new_diseaseType',
    'diseaseClass': 'new_diseaseClass',
    'diseaseSemanticType': 'new_diseaseSemanticType',
    'diseaseId_old': 'diseaseId'
})

merged_df = df.merge(autism_new_ids, on='diseaseId',how='left')
merged_df.head(1)

merged_df['diseaseId'] = merged_df['new_diseaseId'].combine_first(merged_df['diseaseId'])
merged_df['diseaseName'] = merged_df['new_diseaseName'].combine_first(merged_df['diseaseName'])
merged_df['diseaseType'] = merged_df['new_diseaseType'].combine_first(merged_df['diseaseType'])
merged_df['diseaseClass'] = merged_df['new_diseaseClass'].combine_first(merged_df['diseaseClass'])
merged_df['diseaseSemanticType'] = merged_df['new_diseaseSemanticType'].combine_first(merged_df['diseaseSemanticType'])
merged_df.head()

merged_df['autism_subtype_broad'] = merged_df['autism_subtype_broad'].fillna('Not Autism')

grouped_df = merged_df.drop([
    'new_diseaseName',
    'new_diseaseId',
    'new_diseaseType',
    'new_diseaseClass',
    'new_diseaseSemanticType'
    ], axis=1)

grouped_df.head()

grouped_df['autism_subtype_broad'].unique()

# Export data

grouped_df.to_csv('processed_data/curated_gene_disease_associations_autism_grouped.csv', sep='\t')
