==> notebook/4_create_network.ipynb <==
import pandas as pd
import numpy as np
import networkx as nx
import nxviz as nv
from nxviz import MatrixPlot, ArcPlot, CircosPlot
import matplotlib.pyplot as plt

# Import data and delete nodes and edges

nodes = pd.read_csv("processed_data/nodes_diseases_AutismGrouped.csv")
nodes.head(2)

edges = pd.read_csv("processed_data/edges_shared_genes_AutismGrouped.csv")
edges.head(2)

## Delete edges with p > 0.001

edges_p3 = edges[edges['p_log']>=3]
edges_p3.head()

# Load data into NetworkX graph (p<0.01)

# Create the NetworkX graph
G = nx.Graph()

# Add nodes
for idx, row in nodes.iterrows():
    G.add_node(row['diseaseId'], 
               diseaseName=row['diseaseName'],
               diseaseType=row['diseaseType'],
               diseaseClass=row['diseaseClass'],
               diseaseSemanticType=row['diseaseSemanticType'],
               autism_subtype_broad=row['autism_subtype_broad'],
               n_genes=row['n_genes'],
               is_autism=row['is_autism'])

# Add edges
for idx, row in edges_p3.iterrows():
    G.add_edge(row['diseaseId_x'], row['diseaseId_y'], 
               n_shared_genes=row['n_shared_genes'],
               n_genes_x=row['n_genes_x'],
               n_genes_y=row['n_genes_y'],
               n_total_genes=row['n_total_genes'],
               p=row['p'],
               p_log=row['p_log'])

# Degree & Betweenness Centrality

## Degree & Degree centrality

deg_cent = nx.degree_centrality(G)

for n, d in G.nodes(data=True):
    d['degree'] = nx.degree(G, n)
    d['degree_centrality'] = deg_cent[n]

## Betweenness centrality

btw_cent = nx.betweenness_centrality(G)

for n, d in G.nodes(data=True):
    d['btw_centrality'] = btw_cent[n]

## Save network containing all diseases

nx.write_graphml_lxml(G, "graphml/p3_network_all_diseases.graphml")

## Create network only with autism and their neighbors

autism_nodes = [n for n, d in G.nodes(data=True) if d['is_autism']==True]
autism_subgraph_nodes = set(autism_nodes)
for node in autism_nodes:
    autism_subgraph_nodes = autism_subgraph_nodes.union(set(G.neighbors(node)))

autism_subgraph_nodes

G_autism = G.subgraph(autism_subgraph_nodes).copy()

deg_cent = nx.degree_centrality(G_autism)

for n, d in G_autism.nodes(data=True):
    d['degree_autismonly'] = nx.degree(G_autism, n)
    d['degree_centrality_autismonly'] = deg_cent[n]

btw_cent_autism = nx.betweenness_centrality(G_autism)

for n, d in G_autism.nodes(data=True):
    d['btw_centrality_autismonly'] = btw_cent_autism[n]

p_logs = [d['p_log'] for n1, n2, d in G_autism.edges(data=True)]


for n1, n2, d in G_autism.edges(data=True):
    if d['p_log'] >= 10:
        d['p_log_trunc'] = 10
    else:
        d['p_log_trunc'] = int(d['p_log'])

### Save network only with autism and their neighbors

nx.write_graphml_lxml(G_autism, "graphml/p3_network_autism_neighbors.graphml")
