==> notebook/8_autism_disease_class.ipynb <==
import pandas as pd
import numpy as np
import networkx as nx
import nxviz as nv
from nxviz import MatrixPlot, ArcPlot, CircosPlot
import matplotlib.pyplot as plt
import seaborn as sns

data = pd.read_csv("./processed_data/curated_gene_disease_associations_autism_grouped.csv", sep='\t')
autisms = data[data['diseaseId'].str.contains('A')]

autisms = autisms[['diseaseId', 'diseaseName', 'diseaseClass']].drop_duplicates()
autisms

autisms['diseaseClass'] = autisms['diseaseClass'].apply(lambda x: x.split(';'))

autisms

disease_classes = set()
for disease_class in autisms['diseaseClass']:
    disease_classes.update(set(disease_class))
disease_classes = sorted(list(disease_classes))
disease_classes

disease_classes_dicts = [
    ('C04', 'neoplasms'),
    ('C05', 'musculoskeletal diseases'),
    ('C09', 'otorhinolaryngologic diseases'),
    ('C10', 'nervous system diseases'),
    ('C11', 'eye diseases'),
    ('C12', 'urologic and male genital diseases'),
    ('C13', 'female genital diseases and pregnancy complications'),
    ('C14', 'cardiovascular diseases'),
    ('C15', 'hemic and lymphatic diseases'),
    ('C16', 'congenital, hereditary, and neonatal diseases and abnormalities'),
    ('C17', 'skin and connective tissue diseases'),
    ('C18', 'nutritional and metabolic diseases'), 
    ('C19', 'endocrine system diseases'),
    ('C23', 'pathological conditions, signs and symptoms'),
    ('F01', 'behavior and behavior mechanisms'),
    ('F03', 'mental disorders')
]

G = nx.Graph()

# Add class nodes

for class_id, class_name in disease_classes_dicts:
    G.add_node(
        class_id,
        bipartite='disease_class',
        name=class_name)
G.nodes(data=True)

# Add disease nodes

for _, row in autisms.iterrows():
    G.add_node(
        row['diseaseId'],
        bipartite='autism',
        name=row['diseaseName']
        )
list(G.nodes(data=True))[-1]

# Add Edges 

for _, row in autisms.iterrows():
    for c in row['diseaseClass']:
        G.add_edge(row['diseaseId'], c)
G.edges()

# Export

nx.write_graphml_lxml(G, "graphml/bipartite_disease_class.graphml")
